Improve regex for scp syntax that allows special characters
Specials characters like '-' and '_', which are valid in URI's for git
repositories, were getting stripped and resulting in truncated URI
paths.

By slightly modifying the scp regex used in vcs.go, this commit adds
support for these characters and adds a test to ensure that future
changes don't result in a regression.