Made empty user == nil
When no user part of the URL was found, the URL was created with
user == url.User(""). This created a url.String() that was in
the format "scheme://@host/path" which is incompatible with git.

This change makes the user nil when no user part of the URL exists.
This creates url.String() in the format "scheme://host/path".